TASK 4 -- Program  to display contents of array using pointers.
This program first ask the user for the number of elements in the array. Then it takes the input for each element in the array using loops and pointers. Then it displays the contents of the array defined by the user using a 'for' loop and using the same pointer that is used for taking the input.