PROJECT 1: BUS RESERVATION SYSTEM
It's a bus reservation system. It displays bus number, Driver's name, Arrival and departure time, To and From Locations, Driver's Contact number along with Seat availability and also helps to reserve a seat. It also shows the details of the customer who has booked a seat like his/her name, Bus number in which he/she might be travelling from and it also calculates and displays the number of vacant seats after a reservation is made.